wearable_integrations.py
"""
Advanced Wearable API Integrations
Connect to real wearable devices and health platforms
"""

import requests
import json
from datetime import datetime, timedelta
from typing import Dict, List, Any
import logging

logger = logging.getLogger(__name__)

class FitbitIntegration:
    """Fitbit API integration for real health data"""
    
    def __init__(self, client_id: str, client_secret: str):
        self.client_id = client_id
        self.client_secret = client_secret
        self.base_url = "https://api.fitbit.com/1"
        self.access_token = None
    
    def authenticate(self, authorization_code: str) -> bool:
        """OAuth 2.0 authentication with Fitbit"""
        try:
            auth_url = "https://api.fitbit.com/oauth2/token"
            data = {
                'clientId': self.client_id,
                'grant_type': 'authorization_code',
                'redirect_uri': 'http://localhost:5000/callback',
                'code': authorization_code
            }
            
            response = requests.post(auth_url, data=data, auth=(self.client_id, self.client_secret))
            if response.status_code == 200:
                token_data = response.json()
                self.access_token = token_data['access_token']
                return True
            return False
        except Exception as e:
            logger.error(f"Fitbit authentication failed: {e}")
            return False
    
    def get_activity_data(self, date: str) -> Dict[str, Any]:
        """Get daily activity data from Fitbit"""
        if not self.access_token:
            raise ValueError("Not authenticated")
        
        url = f"{self.base_url}/user/-/activities/date/{date}.json"
        headers = {'Authorization': f'Bearer {self.access_token}'}
        
        response = requests.get(url, headers=headers)
        if response.status_code == 200:
            return response.json()
        return {}
    
    def get_sleep_data(self, date: str) -> Dict[str, Any]:
        """Get sleep data from Fitbit"""
        if not self.access_token:
            raise ValueError("Not authenticated")
        
        url = f"{self.base_url}/user/-/sleep/date/{date}.json"
        headers = {'Authorization': f'Bearer {self.access_token}'}
        
        response = requests.get(url, headers=headers)
        if response.status_code == 200:
            return response.json()
        return {}

class AppleHealthIntegration:
    """Apple HealthKit integration"""
    
    def __init__(self):
        self.health_store = None  # Would connect to HealthKit in iOS app
    
    def get_health_data(self, data_types: List[str], start_date: datetime, end_date: datetime) -> Dict[str, Any]:
        """Get health data from Apple Health"""
        # This would be implemented in an iOS app using HealthKit
        # For web, you'd need to use HealthKit Web API or companion app
        pass

class GoogleFitIntegration:
    """Google Fit API integration"""
    
    def __init__(self, credentials_file: str):
        self.credentials_file = credentials_file
        self.service = None
    
    def authenticate(self) -> bool:
        """Authenticate with Google Fit API"""
        try:
            from google.oauth2.credentials import Credentials
            from googleapiclient.discovery import build
            
            # Load credentials and build service
            creds = Credentials.from_authorized_user_file(self.credentials_file)
            self.service = build('fitness', 'v1', credentials=creds)
            return True
        except Exception as e:
            logger.error(f"Google Fit authentication failed: {e}")
            return False
    
    def get_daily_activity(self, date: str) -> Dict[str, Any]:
        """Get daily activity data from Google Fit"""
        if not self.service:
            raise ValueError("Not authenticated")
        
        # Implementation would go here
        pass

class AdvancedDataIntegrator:
    """Advanced data integration with multiple sources"""
    
    def __init__(self):
        self.integrations = {}
        self.user_preferences = {}
    
    def add_integration(self, name: str, integration: Any):
        """Add a new wearable integration"""
        self.integrations[name] = integration
    
    def sync_user_data(self, user_id: str, date_range: tuple) -> Dict[str, Any]:
        """Sync data from all connected sources"""
        combined_data = {
            'user_id': user_id,
            'sync_date': datetime.now().isoformat(),
            'sources': [],
            'metrics': []
        }
        
        for source_name, integration in self.integrations.items():
            try:
                if hasattr(integration, 'get_activity_data'):
                    activity_data = integration.get_activity_data(date_range[0])
                    combined_data['metrics'].append({
                        'source': source_name,
                        'type': 'activity',
                        'data': activity_data
                    })
                
                if hasattr(integration, 'get_sleep_data'):
                    sleep_data = integration.get_sleep_data(date_range[0])
                    combined_data['metrics'].append({
                        'source': source_name,
                        'type': 'sleep',
                        'data': sleep_data
                    })
                
                combined_data['sources'].append(source_name)
                
            except Exception as e:
                logger.error(f"Error syncing {source_name}: {e}")
        
        return combined_data
    
    def get_data_quality_score(self, data: Dict[str, Any]) -> float:
        """Assess data quality and completeness"""
        quality_factors = {
            'completeness': 0.0,
            'consistency': 0.0,
            'freshness': 0.0
        }
        
        # Calculate completeness
        expected_fields = ['steps', 'heart_rate', 'sleep_hours']
        present_fields = sum(1 for field in expected_fields if field in str(data))
        quality_factors['completeness'] = present_fields / len(expected_fields)
        
        # Calculate consistency (data from multiple sources should be similar)
        if len(data.get('sources', [])) > 1:
            quality_factors['consistency'] = 0.8  # Simplified calculation
        
        # Calculate freshness (recent data is better)
        sync_date = datetime.fromisoformat(data.get('sync_date', datetime.now().isoformat()))
        days_old = (datetime.now() - sync_date).days
        quality_factors['freshness'] = max(0, 1 - (days_old / 7))  # Decay over 7 days
        
        return sum(quality_factors.values()) / len(quality_factors)

# Example usage
if __name__ == "__main__":
    # Initialize integrations
    fitbit = FitbitIntegration("your_client_id", "your_client_secret")
    google_fit = GoogleFitIntegration("credentials.json")
    
    # Add to integrator
    integrator = AdvancedDataIntegrator()
    integrator.add_integration("fitbit", fitbit)
    integrator.add_integration("google_fit", google_fit)
    
    # Sync data
    user_data = integrator.sync_user_data("user123", ("2024-11-22", "2024-11-23"))
    quality_score = integrator.get_data_quality_score(user_data)
    
    print(f"Data quality score: {quality_score:.2f}")
