requirements.txt
# Core Web Framework
Flask==3.0.3
Werkzeug==3.0.3
Jinja2==3.1.4
MarkupSafe==2.1.5

# Data Processing & Analysis
numpy==2.0.0
pandas==2.2.2
python-dateutil==2.9.0.post0

# Natural Language Processing
vaderSentiment==3.3.2

# HTTP & API
requests==2.32.3
urllib3==2.2.2

# Utilities
click==8.1.7
itsdangerous==2.2.0
six==1.16.0
pytz==2024.1
tzdata==2024.1

# Optional: Google API Integration (for future wearable API support)
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
oauthlib==3.2.2
requests-oauthlib==2.0.0

# Development & Testing (optional)
pytest==7.4.3
pytest-flask==1.3.0
